requirements.txt
flask==3.0.0
nltk==3.8.1
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.24.3
textblob==0.17.1
rake-nltk==1.0.6
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
python-docx==1.0.1
PyPDF2==3.0.1
textstat==0.7.3
langdetect==1.0.9
gensim==4.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
requests==2.31.0
plotly==5.16.0
